Bench 8563402

+++ src/main.cpp
@@ -664,7 +664,7 @@ i32 alphabeta(Position &pos,
         // minify disable filter delete
 
         i32 score;
-        i32 reduction = (alpha == beta - 1) + !improving - hh_table[pos.flipped][move.from][move.to] / 8192;
+        i32 reduction = (alpha == beta - 1) + !improving - static_cast<i32>(hh_table[pos.flipped][move.from][move.to] / 8192);
         
         if (depth > 2 && num_moves_evaluated > 4 && !gain) {
             reduction = max(-1, min(reduction, depth - 2));
